requirements.txt
requests
psutil
Pillow
tkinterdnd2-universal